requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
chromadb==0.5.0
pdfplumber==0.11.0
python-dotenv==1.0.1
streamlit==1.35.0
python-multipart==0.0.9
tiktoken==0.7.0
